Added a base client
