+++ drc_cmis/browser/client.py
@@ -19,7 +19,7 @@ from drc_cmis.browser.drc_document import (
 )
 from drc_cmis.browser.request import CMISRequest
 from drc_cmis.browser.utils import create_json_request_body
-from drc_cmis.models import Vendor
+from drc_cmis.client import CMISClient
 from drc_cmis.utils.exceptions import (
     CmisUpdateConflictException,
     DocumentConflictException,
@@ -39,12 +39,15 @@ from drc_cmis.utils.utils import build_query_filters, get_random_string
 logger = logging.getLogger(__name__)
 
 
-class CMISDRCClient(CMISRequest):
+class CMISDRCClient(CMISClient, CMISRequest):
     """CMIS client for Browser binding (CMIS 1.1)"""
 
-    _main_repo_id = None
-    _root_folder_id = None
-    _base_folder = None
+    document_type = Document
+    gebruiksrechten_type = Gebruiksrechten
+    oio_type = ObjectInformatieObject
+    folder_type = Folder
+    zaakfolder_type = ZaakFolder
+    zaaktypefolder_type = ZaakTypeFolder
 
     @property
     def base_folder(self) -> Folder:
@@ -73,52 +76,10 @@ class CMISDRCClient(CMISRequest):
 
         return self._root_folder_id
 
-    def get_return_type(self, type_name: str) -> type:
-        error_message = f"No class {type_name} exists for this client."
-        type_name = type_name.lower()
-        assert type_name in [
-            "zaaktypefolder",
-            "zaakfolder",
-            "folder",
-            "document",
-            "gebruiksrechten",
-            "oio",
-        ], error_message
-
-        if type_name == "folder":
-            return Folder
-        elif type_name == "document":
-            return Document
-        elif type_name == "gebruiksrechten":
-            return Gebruiksrechten
-        elif type_name == "oio":
-            return ObjectInformatieObject
-        elif type_name == "zaaktypefolder":
-            return ZaakTypeFolder
-        elif type_name == "zaakfolder":
-            return ZaakFolder
-
     @property
     def vendor(self) -> str:
         repo_info = self.get_request(self.base_url)
         return repo_info["-default-"]["vendorName"]
-
-    def get_object_type_id_prefix(self, object_type: str) -> str:
-        """Get the prefix for the cmis:objectTypeId.
-
-        Alfresco requires prefixes for create statements of custom objects.
-        https://stackoverflow.com/a/28322276/7146757
-
-        :param object_type: str, the type of the object
-        :return: str, the prefix (F: or D:)
-        """
-        if self.vendor.lower() == Vendor.alfresco:
-            if object_type in ["zaaktypefolder", "zaakfolder"]:
-                return "F:"
-            if object_type in ["document", "oio", "gebruiksrechten"]:
-                return "D:"
-
-        return ""
 
     # generic querying
     def query(self, return_type_name: str, lhs: List[str], rhs: List[str]):
@@ -131,67 +92,6 @@ class CMISDRCClient(CMISRequest):
         response = self.post_request(self.base_url, body)
         logger.debug(response)
         return self.get_all_results(response, return_type)
-
-    def get_all_versions(self, document: Document) -> List[Document]:
-        return document.get_all_versions()
-
-    def get_or_create_folder(
-        self, name: str, parent: Folder, properties: dict = None
-    ) -> Folder:
-        """
-        Get or create the folder with :param:`name` in :param:`parent`.
-
-        :param name: string, the name of the folder to create.
-        :param parent: Folder, parent folder to create the folder in as subfolder.
-        :param properties: dict, contains the properties of the folder to create
-        :return: Folder, the folder that was retrieved or created.
-        """
-        logger.debug("CMIS_CLIENT: _get_or_create_folder")
-
-        children_folders = parent.get_children_folders()
-        for folder in children_folders:
-            if folder.name == name:
-                return folder
-
-        return self.create_folder(name, parent.objectId, properties)
-
-    def get_or_create_zaaktype_folder(self, zaaktype: dict) -> Folder:
-        """
-        Create a folder with the prefix 'zaaktype-' to make a zaaktype folder
-
-        Called by ZRC Notification client.
-        """
-        logger.debug("CMIS_CLIENT: get_or_create_zaaktype_folder")
-        properties = {
-            "cmis:objectTypeId": f"{self.get_object_type_id_prefix('zaaktypefolder')}drc:zaaktypefolder",
-            mapper("url", "zaaktype"): zaaktype.get("url"),
-            mapper("identificatie", "zaaktype"): zaaktype.get("identificatie"),
-        }
-
-        folder_name = (
-            f"zaaktype-{zaaktype.get('omschrijving')}-{zaaktype.get('identificatie')}"
-        )
-        cmis_folder = self.get_or_create_folder(
-            folder_name, self.base_folder, properties
-        )
-        return cmis_folder
-
-    def get_or_create_zaak_folder(self, zaak: dict, zaaktype_folder: Folder) -> Folder:
-        """
-        Create a folder with the prefix 'zaak-' to make a zaak folder
-        """
-        logger.debug("CMIS_CLIENT: get_or_create_zaak_folder")
-        properties = {
-            "cmis:objectTypeId": f"{self.get_object_type_id_prefix('zaakfolder')}drc:zaakfolder",
-            mapper("url", "zaak"): zaak.get("url"),
-            mapper("identificatie", "zaak"): zaak.get("identificatie"),
-            mapper("zaaktype", "zaak"): zaak.get("zaaktype"),
-            mapper("bronorganisatie", "zaak"): zaak.get("bronorganisatie"),
-        }
-        cmis_folder = self.get_or_create_folder(
-            f"zaak-{zaak.get('identificatie')}", zaaktype_folder, properties
-        )
-        return cmis_folder
 
     def create_folder(self, name: str, parent_id: str, properties: dict = None):
         logger.debug("CMIS: DRC_DOCUMENT: create_folder")
@@ -229,110 +129,6 @@ class CMISDRCClient(CMISRequest):
         except GetFirstException:
             error_string = f"Folder met objectId 'workspace://SpacesStore/{uuid}' bestaat niet in het CMIS connection"
             raise FolderDoesNotExistError(error_string)
-
-    def delete_cmis_folders_in_base(self):
-        self.base_folder.delete_tree()
-
-    def create_oio(self, data: dict) -> ObjectInformatieObject:
-        """Create ObjectInformatieObject which relates a document with a zaak or besluit
-
-        There are 2 possible cases:
-        1. The document is already related to a zaak: a copy of the document is put in the
-            correct zaaktype/zaak folder.
-        2. The document is not related to anything: the document is moved from the temporary folder
-            to the correct zaaktype/zaak folder.
-
-        If the oio creates a link to a besluit, the zaak/zaaktype need to be retrieved from the besluit.
-
-        If the document is linked already to a gebruiks rechten, then the gebruiksrechten object is also moved.
-
-        :param data: dict, the oio details.
-        :return: Oio created
-        """
-        from drc_cmis.client_builder import get_zds_client
-
-        # Get the document
-        document_uuid = data.get("informatieobject").split("/")[-1]
-        document = self.get_document(uuid=document_uuid)
-
-        if "object" in data:
-            data[data["object_type"]] = data.pop("object")
-
-        # Retrieve the zaak and the zaaktype
-        if data["object_type"] == "besluit":
-            client_besluit = get_zds_client(data["besluit"])
-            besluit_data = client_besluit.retrieve("besluit", url=data["besluit"])
-            zaak_url = besluit_data["zaak"]
-        else:
-            zaak_url = data["zaak"]
-        client_zaak = get_zds_client(zaak_url)
-        zaak_data = client_zaak.retrieve("zaak", url=zaak_url)
-        client_zaaktype = get_zds_client(zaak_data["zaaktype"])
-        zaaktype_data = client_zaaktype.retrieve("zaaktype", url=zaak_data["zaaktype"])
-
-        # Get or create the destination folder
-        zaaktype_folder = self.get_or_create_zaaktype_folder(zaaktype_data)
-        zaak_folder = self.get_or_create_zaak_folder(zaak_data, zaaktype_folder)
-
-        now = timezone.now()
-        year_folder = self.get_or_create_folder(str(now.year), zaak_folder)
-        month_folder = self.get_or_create_folder(str(now.month), year_folder)
-        day_folder = self.get_or_create_folder(str(now.day), month_folder)
-        related_data_folder = self.get_or_create_folder("Related data", day_folder)
-
-        # Check if there are other Oios related to the document
-        retrieved_oios = self.query(
-            return_type_name="oio",
-            lhs=["drc:oio__informatieobject = '%s'"],
-            rhs=[data.get("informatieobject")],
-        )
-
-        # Check if there are gebruiksrechten related to the document
-        related_gebruiksrechten = self.query(
-            return_type_name="gebruiksrechten",
-            lhs=["drc:gebruiksrechten__informatieobject = '%s'"],
-            rhs=[data.get("informatieobject")],
-        )
-
-        # Case 1: Already related to a zaak. Copy the document to the destination folder.
-        if len(retrieved_oios) > 0:
-            self.copy_document(document, day_folder)
-            if len(related_gebruiksrechten) > 0:
-                for gebruiksrechten in related_gebruiksrechten:
-                    self.copy_gebruiksrechten(gebruiksrechten, related_data_folder)
-        # Case 2: Not related to a zaak. Move the document to the destination folder
-        else:
-            document.move_object(day_folder)
-            if len(related_gebruiksrechten) > 0:
-                for gebruiksrechten in related_gebruiksrechten:
-                    gebruiksrechten.move_object(related_data_folder)
-
-        # Create the Oio in a separate folder
-        return self.create_content_object(
-            data=data, object_type="oio", destination_folder=related_data_folder
-        )
-
-    def create_gebruiksrechten(self, data: dict) -> Gebruiksrechten:
-        """Create gebruiksrechten
-
-        The geburiksrechten is created in the 'Related data' folder in the folder
-        of the related document
-
-        :param data: dict, data of the gebruiksrechten
-        :return: Gebruiksrechten
-        """
-
-        document_uuid = data.get("informatieobject").split("/")[-1]
-        document = self.get_document(uuid=document_uuid)
-
-        parent_folder = document.get_parent_folders()[0]
-        related_data_folder = self.get_or_create_folder("Related data", parent_folder)
-
-        return self.create_content_object(
-            data=data,
-            object_type="gebruiksrechten",
-            destination_folder=related_data_folder,
-        )
 
     def copy_gebruiksrechten(
         self, source_object: Gebruiksrechten, destination_folder: Folder
@@ -501,24 +297,8 @@ class CMISDRCClient(CMISRequest):
             )
             raise DocumentDoesNotExistError(error_string)
 
-    def delete_content_object(self, uuid: Union[str, UUID], object_type: str):
-        """Delete the gebruiksrechten/objectinformatieobject with specified uuid
-
-        :param uuid: string or UUID, identifier that when combined with 'workspace://SpacesStore/' and the version
-        number gives the cmis:objectId
-        :param object_type: string, either "gebruiksrechten" or "oio"
-        :return: Either a Gebruiksrechten or ObjectInformatieObject
-        """
-
-        content_object = self.get_content_object(uuid, object_type=object_type)
-        content_object.delete_object()
-
     def create_document(
-        self,
-        identification: str,
-        data: dict,
-        target_folder: Folder = None,
-        content: BytesIO = None,
+        self, identification: str, data: dict, content: BytesIO = None,
     ) -> Document:
         """Create a cmis document.
 
@@ -540,16 +320,16 @@ class CMISDRCClient(CMISRequest):
         if content is None:
             content = BytesIO()
 
-        if target_folder is None:
-            year_folder = self.get_or_create_folder(str(now.year), self.base_folder)
-            month_folder = self.get_or_create_folder(str(now.month), year_folder)
-            target_folder = self.get_or_create_folder(str(now.day), month_folder)
+        # Create Document in default folder
+        year_folder = self.get_or_create_folder(str(now.year), self.base_folder)
+        month_folder = self.get_or_create_folder(str(now.month), year_folder)
+        day_folder = self.get_or_create_folder(str(now.day), month_folder)
 
         properties = Document.build_properties(
             data, new=True, identification=identification
         )
 
-        data = create_json_request_body(target_folder, properties)
+        data = create_json_request_body(day_folder, properties)
 
         json_response = self.post_request(self.root_folder_url, data=data)
         cmis_doc = Document(json_response)
@@ -679,11 +459,6 @@ class CMISDRCClient(CMISRequest):
             return self.get_first_result(json_response, Document)
         except GetFirstException as exc:
             raise does_not_exist from exc
-
-    def delete_document(self, uuid: str) -> None:
-        """Delete all versions of a document with objectId workspace://SpacesStore/<uuid>"""
-        document = self.get_document(uuid=uuid)
-        document.delete_object()
 
     def check_document_exists(self, identification: Union[str, UUID]):
         """Query by identification (``identificatie``) if a document is in the repository"""
